Move everything into Version and make button display currently selected version
