+++ pom.xml
@@ -21,6 +21,7 @@
     </repositories>
 
     <dependencies>
+        <!-- General -->
         <dependency>
             <groupId>me.grison</groupId>
             <artifactId>jtoml</artifactId>
@@ -32,10 +33,25 @@
             <version>1.6</version>
         </dependency>
         <dependency>
+            <groupId>org.jetbrains</groupId>
+            <artifactId>annotations</artifactId>
+            <version>RELEASE</version>
+            <scope>compile</scope>
+        </dependency>
+        <!-- General End -->
+
+        <!-- Data -->
+        <dependency>
             <groupId>com.google.code.gson</groupId>
             <artifactId>gson</artifactId>
             <version>2.8.8</version>
         </dependency>
+        <dependency>
+            <groupId>org.rauschig</groupId>
+            <artifactId>jarchivelib</artifactId>
+            <version>0.7.1</version>
+        </dependency>
+        <!-- Data End -->
 
         <!-- JavaFX -->
         <dependency>
@@ -74,39 +90,17 @@
             <version>17.0.0.1</version>
             <type>pom</type>
         </dependency>
-        <!-- JavaFX End -->
-
-        <dependency>
-            <groupId>org.jetbrains</groupId>
-            <artifactId>annotations</artifactId>
-            <version>RELEASE</version>
-            <scope>compile</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.kordamp.ikonli</groupId>
-            <artifactId>ikonli-bootstrapicons-pack</artifactId>
-            <version>12.2.0</version>
-        </dependency>
-
         <dependency>
             <groupId>org.kordamp.ikonli</groupId>
             <artifactId>ikonli-javafx</artifactId>
             <version>12.2.0</version>
         </dependency>
-
         <dependency>
-            <groupId>net.sf.jopt-simple</groupId>
-            <artifactId>jopt-simple</artifactId>
-            <version>6.0-alpha-3</version>
+            <groupId>org.kordamp.ikonli</groupId>
+            <artifactId>ikonli-bootstrapicons-pack</artifactId>
+            <version>12.2.0</version>
         </dependency>
-
-        <dependency>
-            <groupId>org.rauschig</groupId>
-            <artifactId>jarchivelib</artifactId>
-            <version>0.7.1</version>
-        </dependency>
-
+        <!-- JavaFX End -->
     </dependencies>
